polish backend function modularity 1
